+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- In `DevConsole.CreatePanel` check no panel has the same name as the one being added.
 
 ### Changed
 - Replaced `SetPreActive` and `SetActive` with `PreActivate`, `PreDeactivate`, `Activate` and `Deactivate`.